PR changes request complete

+++ models/migrations/20191004135959_create-free-token-column.js
@@ -1,6 +1,6 @@
 exports.up = function(knex) {
   return knex.schema.alterTable('users', (tbl) => {
-    tbl.bigInteger('credits').notNullable().defaultTo(1);
+    tbl.integer('credits').notNullable().defaultTo(1);
   });
 };
